+++ lib/flutter_countdown.dart
@@ -5,29 +5,43 @@ import 'package:flutter/cupertino.dart';
 
 class _Countdown extends AnimatedWidget {
   final Animation<int> animation;
-  final String Function(int number) renderSemanticLabel;
+  final String Function(int number) renderLabel;
   final TextStyle textStyle;
 
-  _Countdown({Key key, this.animation, this.renderSemanticLabel, this.textStyle})
+  _Countdown(
+      {Key key, this.animation, this.renderLabel, this.textStyle})
       : super(key: key, listenable: animation);
 
   @override
   build(BuildContext context) {
     int count = animation.value;
-    return Text(renderSemanticLabel(count), style: textStyle, textDirection: TextDirection.ltr,);
+    return Text(
+      renderLabel(count),
+      style: textStyle,
+      textDirection: TextDirection.ltr,
+    );
   }
 }
 
 class CountDown extends StatefulWidget {
   final int beginCount;
   final int endCount;
+  // renderSemanticLabel is deprecated and shouldn't be used.
+  // you can use renderLabel
   final String Function(int number) renderSemanticLabel;
+  final String Function(int number) renderLabel;
   final void Function(AnimationController refs) refs;
   final Future<bool> Function(AnimationController ctr) onPress;
   final AnimationStatusListener statusListener;
 
-  CountDown({this.beginCount,
-      this.endCount, this.statusListener, this.refs, this.onPress, this.renderSemanticLabel});
+  CountDown(
+      {this.beginCount,
+      this.endCount,
+      this.statusListener,
+      this.refs,
+      this.onPress,
+      this.renderLabel,
+      this.renderSemanticLabel});
 
   @override
   State<StatefulWidget> createState() => CounDownState();
@@ -55,8 +69,8 @@ class CounDownState extends State<CountDown> with TickerProviderStateMixin {
     });
 
     if (widget.statusListener != null) {
-        widget.refs(_controller);
-      }
+      widget.refs(_controller);
+    }
   }
 
   _beginCountIfNeed(AnimationController ctr) async {
@@ -67,10 +81,10 @@ class CounDownState extends State<CountDown> with TickerProviderStateMixin {
   }
 
   @override
-    void dispose() {
-      _controller.dispose();
-      super.dispose();
-    }
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -78,7 +92,13 @@ class CounDownState extends State<CountDown> with TickerProviderStateMixin {
       onPressed: () => _beginCountIfNeed(_controller),
       padding: const EdgeInsets.all(0),
       child: _Countdown(
-        renderSemanticLabel: (count) => widget.renderSemanticLabel(count),
+        renderLabel: (count) {
+          if (widget.renderLabel != null) {
+            widget.renderLabel(count);
+            return;
+          }
+          widget.renderSemanticLabel(count);
+        },
         animation: new StepTween(
           begin: widget.beginCount,
           end: widget.endCount,