+++ CHANGELOG.md
@@ -5,3 +5,12 @@
 ## [0.1.1] - 2018-11-16 10:37:25
 
 * remove button padding
+
+## [0.1.3] - 2018-11-16 10:37:25
+
+* Increase package's point
+
+## [0.1.4] - 2018-11-16 10:37:25
+
+* renderSemanticLabel => renderLabel
+* renderSemanticLabel is deprecated and shouldn't be used.